Add 'work-in-progress' identifier to grammar
Increasingly so there are changes in the codebase that may not be brittle, but would still not qualify as a _"Feature"_. This PR asks for the addition of "WIP" designator to the language-todo as a first-class citizen. This would be useful in the environment where a number of daily commits is necessary to facilitate team collaboration and peer review, as well as certain product- and project-management KPIs.